NMR-6922: extract chart synchronization in a separate object
