+++ nmrfx-analyst-gui/src/main/java/org/nmrfx/processor/gui/AxisChangeListener.java
@@ -68,17 +68,17 @@ public class AxisChangeListener implements ChangeListener<Number> {
                     chart.controller.getStatusBar().updatePlaneSpinner(indexL, axNum, 0);
                     chart.controller.getStatusBar().updatePlaneSpinner(indexU, axNum, 1);
                 }
-                if (PolyChart.getNSyncGroups() > 0) {
+                if (PolyChartSynchronizer.getNSyncGroups() > 0) {
                     List<String> names = chart.getDimNames();
                     String name = names.get(axNum);
-                    int syncGroup = chart.getSyncGroup(name);
+                    int syncGroup = chart.getSynchronizer().getSyncGroup(name);
 
                     PolyChartManager.getInstance().getAllCharts().stream().filter((otherChart) -> (otherChart != chart)).forEach((otherChart) -> {
                         List<String> otherNames = otherChart.getDimNames();
                         int i = 0;
                         for (String otherName : otherNames) {
                             if (otherName.equals(name)) {
-                                int otherGroup = otherChart.getSyncGroup(otherName);
+                                int otherGroup = otherChart.getSynchronizer().getSyncGroup(otherName);
                                 if ((otherGroup > 0) && (syncGroup == otherGroup)) {
                                     if (endNum == 0) {
                                         otherChart.axes[i].setLowerBound(newBound);
